+++ src/com/intellij/idea/plugin/hybris/common/utils/HybrisIcons.kt
@@ -240,6 +240,7 @@ object HybrisIcons {
     val CCV2_BUILD_REDO = AllIcons.Actions.BuildAutoReloadChanges
     val CCV2_BUILD_DEPLOY = AllIcons.Nodes.Deploy
     val CCV2_BUILD_DELETE = ExpUiIcons.General.Delete
+    val CCV2_BUILD_LOGS = ExpUiIcons.General.Download
 
     val CCV2_DEPLOYMENT_CREATED_BY = AllIcons.General.User
     val CCV2_DEPLOYMENT_UPDATE_MODE_NONE = AllIcons.Diff.GutterCheckBox