+++ src/com/intellij/idea/plugin/hybris/toolwindow/ccv2/views/CCv2BuildsDataView.kt
@@ -22,6 +22,7 @@ import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.settings.CCv2Subscription
 import com.intellij.idea.plugin.hybris.tools.ccv2.actions.CCv2DeleteBuildAction
 import com.intellij.idea.plugin.hybris.tools.ccv2.actions.CCv2DeployBuildAction
+import com.intellij.idea.plugin.hybris.tools.ccv2.actions.CCv2DownloadBuildLogsAction
 import com.intellij.idea.plugin.hybris.tools.ccv2.actions.CCv2RedoBuildAction
 import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2Build
 import com.intellij.idea.plugin.hybris.toolwindow.ccv2.CCv2Tab
@@ -58,6 +59,7 @@ object CCv2BuildsDataView : AbstractCCv2DataView<CCv2Build>() {
                             CCv2RedoBuildAction(subscription, build),
                             if (build.canDeploy()) CCv2DeployBuildAction(subscription, build) else null,
                             if (build.canDelete()) CCv2DeleteBuildAction(subscription, build) else null,
+                            if (build.canDownloadLogs()) CCv2DownloadBuildLogsAction(subscription, build) else null,
                         ).toTypedArray(),
                         ActionPlaces.TOOLWINDOW_CONTENT
                     )